TAO-10285 Moved query execution to single method and exceptions to propper namespace
